sync original

+++ submodules/LegacyComponents/Sources/TGMediaOriginInfo.m
@@ -275,7 +275,7 @@
     return info;
 }
 
-+ (instancetype)mediaOriginInfoWithFileReference:(NSData *)fileReference fileReferences:(NSDictionary *)fileReferences userId:(int32_t)userId offset:(int32_t)offset
++ (instancetype)mediaOriginInfoWithFileReference:(NSData *)fileReference fileReferences:(NSDictionary *)fileReferences userId:(int64_t)userId offset:(int32_t)offset
 {
     TGMediaOriginInfo *info = [[TGMediaOriginInfo alloc] init];
     info->_type = TGMediaOriginTypeProfilePhoto;
